[MIRROR] [MDB IGNORE] Restore non-functional ash mark ability (#9675)
Original PR: https://github.com/tgstation/tgstation/pull/81320
-----

## About The Pull Request

Apparently, ash mark is supposed to replenish mansus grasp cooldown by
75% when triggered, but this is neither reflected by the research
description, nor functional ingame. This edit makes it work, and adds
the effect to the description.

Additionally, that same block of code was present in moon mark research
also. I removed that because I assume it is intended to be an ash-only
feature, and moon only had it because of copy-pasting.

And since we're at it, I also reworded the mark description to be more
clear about the mark only transferring to a single heathen at a time.
## Why It's Good For The Game

I don't think we should have code which does not work. The options are
to either remove it, or make it work, and the latter is more fun.
## Changelog
:cl:
fix: ash mark now properly lowers the cooldown of mansus grasp when
triggered
/:cl:

---------

Co-authored-by: ViktorKoL <44502667+ViktorKoL@users.noreply.github.com>