[MIRROR] [MDB IGNORE] Fixes the typepath of the shove blocker module. (#9719)
Original PR: https://github.com/tgstation/tgstation/pull/81374
-----

## About The Pull Request

The bulwark module had the wrong typepath for most of the descriptive
elements and its complexity. So the shove block was a free module. And
technically not incompatible with itself. Oh my.

Not super relevant for actual play, as there is no access to this module
anywhere currently, but who knows.

## Why It's Good For The Game

Typepaths.

## Changelog
:cl:
fix: The shove blocker module parent type now has the correct typepath.
/:cl:

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>