Fixes Cargo Lockbox Iconstates (#81481)
## About The Pull Request

The cargo lockboxes didn't update iconstates when locked. Seems like an
oversight given that their parent object does this and there's unused
sprites for this behaviour in this item, so this enables it to use those
sprites that are already in code.

## Why It's Good For The Game

It's hard to know if someone has taken their items or not because an
unlocked case looks identical to a locked one. Also fixes an oversight
idk.

## Changelog
:cl:
fix: cargo lockboxes update iconstates correctly now
/:cl: