Removes Skrell from the Game (#9862)
I've been teetering on this for months, years even. Lore team is in
agreeance that Skrell is to be removed.

Also they really don't fit in with what we want to do.


![image](https://github.com/user-attachments/assets/97d0db54-07fa-4ecc-90b2-ca8d0da74783)

RIP Skrell, we added you over 3 years ago for one person to play as you
for one round. You survived long enough to be the backbone of many
species to come after you. But frankly, you're weird. Goodbye.

## Changelog

:cl: Jolly
del: Skrell has been removed. Warble warble.
/:cl: